+++ packages/backend/src/models/models.ts
@@ -11,6 +11,7 @@ import { GithubAppInstallationsModel } from './GithubAppInstallations/GithubAppI
 import { GroupsModel } from './GroupsModel';
 import { InviteLinkModel } from './InviteLinkModel';
 import { JobModel } from './JobModel/JobModel';
+import { NotificationsModel } from './NotificationsModel/NotificationsModel';
 import { OnboardingModel } from './OnboardingModel/OnboardingModel';
 import { OpenIdIdentityModel } from './OpenIdIdentitiesModel';
 import { OrganizationAllowedEmailDomainsModel } from './OrganizationAllowedEmailDomainsModel';
@@ -35,9 +36,12 @@ import { ValidationModel } from './ValidationModel/ValidationModel';
 
 export const encryptionService = new EncryptionService({ lightdashConfig });
 
-export const inviteLinkModel = new InviteLinkModel(database);
+export const inviteLinkModel = new InviteLinkModel({
+    database,
+    lightdashConfig,
+});
 export const organizationModel = new OrganizationModel(database);
-export const userModel = new UserModel(database);
+export const userModel = new UserModel({ database, lightdashConfig });
 export const sessionModel = new SessionModel(database);
 export const dashboardModel = new DashboardModel({ database });
 export const projectModel = new ProjectModel({
@@ -115,3 +119,5 @@ export const githubAppInstallationsModel = new GithubAppInstallationsModel({
 });
 
 export const commentModel = new CommentModel({ database });
+
+export const notificationsModel = new NotificationsModel({ database });