+++ packages/common/src/types/featureFlags.ts
@@ -6,12 +6,6 @@
  */
 export enum FeatureFlags {
     /**
-     * Use dbt ls when compiling lightdash projects with "refresh dbt"
-     * See  ProjectService for example usage.
-     */
-    UseDbtLs = 'use-dbt-ls',
-
-    /**
      * Use shared color assignments for cartesian-type charts, when possible. This
      * essentially means we try to have the same dimension translate into the same
      * color in the org's color palette every time.
@@ -33,11 +27,6 @@ export enum FeatureFlags {
      * Use the new in-memory table calculations engine/duckdb
      */
     UseInMemoryTableCalculations = 'new-table-calculations-engine',
-
-    /**
-     * Enables the new omnibar search
-     */
-    Omnibar = 'omnibar',
 
     /**
      * Use the new dashboard comments feature. Users can add/resolve/delete comments