+++ packages/common/src/types/search.ts
@@ -1,3 +1,4 @@
+import assertUnreachable from '../utils/assertUnreachable';
 import { Dashboard } from './dashboard';
 import { Table } from './explore';
 import { Dimension, Metric } from './field';
@@ -40,6 +41,7 @@ export type TableSearchResult = Pick<
 > & {
     explore: string;
     exploreLabel: string;
+    regexMatchCount: number;
 };
 
 export type TableErrorSearchResult = Pick<
@@ -68,6 +70,7 @@ export type FieldSearchResult = Pick<
     >;
     explore: string;
     exploreLabel: string;
+    regexMatchCount: number;
 };
 
 type PageResult = {
@@ -127,6 +130,30 @@ export enum SearchItemType {
     TABLE = 'table',
     FIELD = 'field',
     PAGE = 'page',
+}
+
+export function getSearchItemTypeFromResultKey(
+    searchResultKey: keyof SearchResults,
+) {
+    switch (searchResultKey) {
+        case 'spaces':
+            return SearchItemType.SPACE;
+        case 'dashboards':
+            return SearchItemType.DASHBOARD;
+        case 'savedCharts':
+            return SearchItemType.CHART;
+        case 'tables':
+            return SearchItemType.TABLE;
+        case 'fields':
+            return SearchItemType.FIELD;
+        case 'pages':
+            return SearchItemType.PAGE;
+        default:
+            return assertUnreachable(
+                searchResultKey,
+                `unexpected search result key: ${searchResultKey}`,
+            );
+    }
 }
 
 export type SearchFilters = {