+++ packages/frontend/src/features/omnibar/components/Omnibar.tsx
@@ -1,7 +1,9 @@
 import {
+    getSearchItemTypeFromResultKey,
     getSearchResultId,
     SearchFilters,
     SearchItemType,
+    SearchResults,
 } from '@lightdash/common';
 import {
     ActionIcon,
@@ -14,9 +16,14 @@ import {
     Transition,
     useMantineTheme,
 } from '@mantine/core';
-import { useDebouncedValue, useDisclosure, useHotkeys } from '@mantine/hooks';
+import {
+    useDebouncedValue,
+    useDisclosure,
+    useHotkeys,
+    useScrollIntoView,
+} from '@mantine/hooks';
 import { IconCircleXFilled, IconSearch } from '@tabler/icons-react';
-import { FC, MouseEventHandler, useState } from 'react';
+import { FC, MouseEventHandler, useEffect, useMemo, useState } from 'react';
 import { useHistory, useLocation } from 'react-router-dom';
 import MantineIcon from '../../../components/common/MantineIcon';
 import { PAGE_CONTENT_WIDTH } from '../../../components/common/Page/Page';
@@ -25,11 +32,16 @@ import { useValidationUserAbility } from '../../../hooks/validation/useValidatio
 import { useTracking } from '../../../providers/TrackingProvider';
 import { EventName } from '../../../types/Events';
 import useSearch, { OMNIBAR_MIN_QUERY_LENGTH } from '../hooks/useSearch';
-import { allSearchItemTypes, SearchItem } from '../types/searchItem';
+import {
+    allSearchItemTypes,
+    FocusedItemIndex,
+    SearchItem,
+} from '../types/searchItem';
 import { isSearchResultEmpty } from '../utils/isSearchResultEmpty';
 import OmnibarEmptyState from './OmnibarEmptyState';
 import OmnibarFilters from './OmnibarFilters';
 import OmnibarItemGroups from './OmnibarItemGroups';
+import { OmnibarKeyboardNav } from './OmnibarKeyboardNav';
 import OmnibarTarget from './OmnibarTarget';
 
 interface Props {
@@ -42,15 +54,16 @@ const Omnibar: FC<Props> = ({ projectUuid }) => {
     const { data: projectData } = useProject(projectUuid);
     const { track } = useTracking();
     const theme = useMantineTheme();
-
     const canUserManageValidation = useValidationUserAbility(projectUuid);
-
-    const [openPanels, setOpenPanels] =
-        useState<SearchItemType[]>(allSearchItemTypes);
-
     const [searchFilters, setSearchFilters] = useState<SearchFilters>();
     const [query, setQuery] = useState<string>();
     const [debouncedValue] = useDebouncedValue(query, 300);
+    const { targetRef: scrollRef } = useScrollIntoView<HTMLDivElement>(); // couldn't get scroll to work with mantine's function
+    const [openPanels, setOpenPanels] =
+        useState<SearchItemType[]>(allSearchItemTypes);
+
+    const [focusedItemIndex, setFocusedItemIndex] =
+        useState<FocusedItemIndex>();
 
     const { data: searchResults, isFetching } = useSearch(
         projectUuid,
@@ -99,6 +112,7 @@ const Omnibar: FC<Props> = ({ projectUuid }) => {
             },
         });
 
+        setFocusedItemIndex(undefined);
         closeOmnibar();
     };
 
@@ -136,8 +150,37 @@ const Omnibar: FC<Props> = ({ projectUuid }) => {
     const hasSearchResults =
         searchResults && !isSearchResultEmpty(searchResults);
 
+    const sortedGroupEntries = useMemo(() => {
+        return searchResults
+            ? Object.entries(searchResults)
+                  .map((items) => {
+                      return [
+                          getSearchItemTypeFromResultKey(
+                              items[0] as keyof SearchResults,
+                          ),
+                          items[1],
+                      ] as [SearchItemType, SearchItem[]];
+                  })
+                  .filter(([_type, items]) => items.length > 0)
+                  .sort(
+                      ([_a, itemsA], [_b, itemsB]) =>
+                          (itemsB[0].searchRank ?? 0) -
+                          (itemsA[0].searchRank ?? 0),
+                  )
+            : [];
+    }, [searchResults]);
+
+    useEffect(() => {
+        setFocusedItemIndex(undefined);
+    }, [query, searchFilters]);
+
     return (
-        <>
+        <OmnibarKeyboardNav
+            groupedItems={sortedGroupEntries}
+            onEnterPressed={handleItemClick}
+            onFocusedItemChange={setFocusedItemIndex}
+            currentFocusedItemIndex={focusedItemIndex}
+        >
             <Transition
                 mounted={!isOmnibarOpen}
                 transition="fade"
@@ -242,7 +285,6 @@ const Omnibar: FC<Props> = ({ projectUuid }) => {
                             <OmnibarEmptyState message="No results found." />
                         ) : (
                             <OmnibarItemGroups
-                                searchResults={searchResults}
                                 projectUuid={projectUuid}
                                 canUserManageValidation={
                                     canUserManageValidation
@@ -250,12 +292,15 @@ const Omnibar: FC<Props> = ({ projectUuid }) => {
                                 openPanels={openPanels}
                                 onOpenPanelsChange={setOpenPanels}
                                 onClick={handleItemClick}
+                                focusedItemIndex={focusedItemIndex}
+                                groupedItems={sortedGroupEntries}
+                                scrollRef={scrollRef}
                             />
                         )}
                     </Stack>
                 </Modal>
             </MantineProvider>
-        </>
+        </OmnibarKeyboardNav>
     );
 };